+++ test/manifolds/multinomial_matrices.jl
@@ -55,4 +55,12 @@ include("../header.jl")
         M = MultinomialMatrices(3, 2; parameter=:field)
         @test repr(M) == "MultinomialMatrices(3, 2; parameter=:field)"
     end
+    @testset "Riemannian Gradient" begin
+        M = MultinomialMatrices(3, 2)
+        p = [0.5 0.4 0.1; 0.5 0.4 0.1]'
+        Y = [1.0 -1.0; 0.0 0.0; -1.0 1.0]
+        G = project(M, p, p .* Y)
+        X = riemannian_gradient(M, p, Y)
+        @test isapprox(M, p, G, X)
+    end
 end